@@ -12,7 +12,7 @@
   "dependencies": {
     "@fastentregas/axios-config": "*",
     "@hookform/resolvers": "^2.8.8",
-    "@react-native-async-storage/async-storage": "^1.17.3",
+    "@react-native-async-storage/async-storage": "~1.15.0",
     "@react-navigation/material-top-tabs": "^6.2.1",
     "@react-navigation/native": "^6.0.10",
     "@react-navigation/stack": "^6.2.1",
@@ -22,7 +22,7 @@
     "expo-splash-screen": "~0.14.1",
     "expo-status-bar": "~1.2.0",
     "expo-yarn-workspaces": "^1.7.0",
-    "react": "~17.0.0",
+    "react": "^18",
     "react-hook-form": "^7.30.0",
     "react-native": "0.64.3",
     "react-native-gesture-handler": "~2.1.0",
@@ -32,7 +32,7 @@
     "react-native-screens": "~3.10.1",
     "react-native-tab-view": "~3.1.1",
     "react-native-web": "0.17.1",
-    "styled-components": "^5.3.5",
+    "styled-components": "^6",
     "yup": "^0.32.11"
   },
   "devDependencies": {
