@@ -9,26 +9,28 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@emotion/cache": "^11.11.0",
+    "@emotion/react": "^11.11.3",
     "@fastentregas/axios-config": "*",
     "@hookform/resolvers": "^2.8.8",
-    "@mui/icons-material": "^5.6.2",
-    "@mui/material": "^5.6.2",
-    "@mui/styled-engine-sc": "~5.8.0",
+    "@mui/icons-material": "^5.15.3",
+    "@mui/material": "^5.15.3",
+    "@mui/material-nextjs": "^5.15.3",
+    "@mui/styled-engine-sc": "~6.0.0-alpha.11",
     "@types/cookie": "^0.5.1",
-    "next": "12.1.5",
-    "next-transpile-modules": "^9.0.0",
+    "next": "~14.0.0",
+    "next-transpile-modules": "^10.0.0",
     "nookies": "^2.5.2",
-    "react": "~17.0.0",
-    "react-dom": "~17.0.0",
+    "react": "^18",
+    "react-dom": "^18",
     "react-hook-form": "^7.30.0",
     "react-input-mask": "^3.0.0-alpha.2",
-    "styled-components": "^5.3.5",
+    "styled-components": "^6",
     "tp-react-web-masked-text": "^1.12.22",
     "yup": "^0.32.11"
   },
   "devDependencies": {
     "@types/react": "^17.0.21",
-    "@types/react-input-mask": "^3.0.1",
-    "@types/styled-components": "^5.1.25"
+    "@types/react-input-mask": "^3.0.1"
   }
 }
