@@ -1,10 +1,12 @@
 import React from 'react';
-import AppBar from '@mui/material/AppBar';
-import Moped from '@mui/icons-material/Moped';
-import Container from '@mui/material/Container';
-import Typography from '@mui/material/Typography';
-import Toolbar from '@mui/material/Toolbar';
-import CssBaseline from '@mui/material/CssBaseline';
+import {
+  AppBar,
+  CssBaseline,
+  Toolbar,
+  Container,
+  Typography,
+} from '@mui/material';
+import { Moped } from '@mui/icons-material';
 
 export default function About(): React.ReactElement {
   return (
